chore: fully split aws related modules from services package
