+++ app/infrastructure/storage/__init__.py
@@ -13,5 +13,6 @@ Usage::
 """
 
 from infrastructure.storage.protocol import StorageService
+from infrastructure.storage.service import get_storage_service
 
-__all__ = ["StorageService"]
+__all__ = ["StorageService", "get_storage_service"]
